policy gate

+++ app/Http/Controllers/ArticleController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Article;
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+
 
 class ArticleController extends Controller
 {
@@ -26,6 +28,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', [self::class]);
         return view('article.create');
     }
 
@@ -37,6 +40,8 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', [self::class]);
+
         $request->validate([
             'date' => 'date',
             'title' => 'required|max:20|min:3',
@@ -73,6 +78,8 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('create', [self::class]);
+
         return view('article.edit', ['article'=>$article]);
     }
 
@@ -85,6 +92,8 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        Gate::authorize('create', [self::class]);
+
         $request->validate([
             'date' => 'date',
             'title' => 'required|max:20|min:3',
@@ -107,6 +116,8 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('create', [self::class]);
+
         Comment::where('article_id', $article->id)->delete();
         $article->delete();
         return redirect()->route('article.index');
