+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
+
 
 class CommentController extends Controller
 {
@@ -23,6 +25,7 @@ class CommentController extends Controller
 
     public function edit($id){
         $comment = Comment::findOrFail($id);
+        Gate::authorize('comment', $comment);
         return view('comment.edit',['comment'=>$comment] );
     }
     public function update(Request $request, $id){
@@ -31,13 +34,16 @@ class CommentController extends Controller
         ]);
 
         $comment = Comment::findOrFail($id);
+        Gate::authorize('comment', $comment);
         $comment->title = $request->title;
         $comment->text = $request->text;
         $comment->save();
         return redirect()->route('article.show', ['article'=>$request->article_id]);
     }
     public function delete($id){
+        
         $comment = Comment::findOrFail($id);
+        Gate::authorize('comment', $comment);
         $comment->delete();
         // $comment = NULL;
         return redirect()->route('article.show', ['article'=>$comment->article_id]);
